Book List: Delete confirmation box
A box wil confirmation box will appear when trying to delete a book
from the list. Also deleting the book on the list will also delete it's
image from the storage. Makes sure books aren't deleted by mistake also
ensure unsused images are deleted, to keep storage clean

Closes #48

+++ src/app/books-list/books-list.component.ts
@@ -21,7 +21,9 @@ export class BooksListComponent implements OnInit {
   }
 
   deleteBook(data) { 
-    this.bookStorage.deleteBook(data);
+    if(confirm("Are you sure you want to delete this book?")){
+      this.bookStorage.deleteBook(data);
+    }
   }
 
   saveBookOnClick(book: Book) {
